Compiler: move macrotable to preprocessor sources
