+++ Solutions/Compiler.Lib/Compiler.Lib.vcxproj
@@ -297,7 +297,7 @@
     <ClCompile Include="..\..\Compiler\fmem.cpp" />
     <ClCompile Include="..\..\Compiler\script\cc_compiledscript.cpp" />
     <ClCompile Include="..\..\Compiler\script\cc_internallist.cpp" />
-    <ClCompile Include="..\..\Compiler\script\cc_macrotable.cpp" />
+    <ClCompile Include="..\..\Compiler\preproc\cc_macrotable.cpp" />
     <ClCompile Include="..\..\Compiler\script\cc_symboltable.cpp" />
     <ClCompile Include="..\..\Compiler\script\cc_treemap.cpp" />
     <ClCompile Include="..\..\Compiler\script\cs_compiler.cpp" />
@@ -316,7 +316,7 @@
     <ClInclude Include="..\..\Compiler\fmem.h" />
     <ClInclude Include="..\..\Compiler\script\cc_compiledscript.h" />
     <ClInclude Include="..\..\Compiler\script\cc_internallist.h" />
-    <ClInclude Include="..\..\Compiler\script\cc_macrotable.h" />
+    <ClInclude Include="..\..\Compiler\preproc\cc_macrotable.h" />
     <ClInclude Include="..\..\Compiler\script\cc_symboldef.h" />
     <ClInclude Include="..\..\Compiler\script\cc_symboltable.h" />
     <ClInclude Include="..\..\Compiler\script\cc_treemap.h" />
