+++ Solutions/Compiler.Lib/Compiler.Lib.vcxproj.filters
@@ -41,7 +41,7 @@
     <ClCompile Include="..\..\Compiler\script\cc_internallist.cpp">
       <Filter>Source Files\script</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\Compiler\script\cc_macrotable.cpp">
+    <ClCompile Include="..\..\Compiler\preproc\cc_macrotable.cpp">
       <Filter>Source Files\script</Filter>
     </ClCompile>
     <ClCompile Include="..\..\Compiler\script\cc_symboltable.cpp">
@@ -104,7 +104,7 @@
     <ClInclude Include="..\..\Compiler\script\cc_internallist.h">
       <Filter>Header Files\script</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\Compiler\script\cc_macrotable.h">
+    <ClInclude Include="..\..\Compiler\preproc\cc_macrotable.h">
       <Filter>Header Files\script</Filter>
     </ClInclude>
     <ClInclude Include="..\..\Compiler\script\cc_symboldef.h">
